requirements.txt
PyQt5>=5.15.0
tinkoff-investments>=0.2.0
protobuf>=3.20.0
asyncio-mqtt>=0.11.0 